+++ adam_sim/features/base/base.py
@@ -8,8 +8,10 @@ class Base:
     TMP
     '''
 
-    def __init__(self, id: str) -> None:
+    def _init_repository(self, id: str, *args, **kwargs) -> None:
         self._repository: BaseRepository = get_base_repository(id)
+
+        self._repository.init(*args, **kwargs)
 
     def move_to(self, position: Point, orientation: Vector) -> None:
         pass